refactor(queue): extract review-evasion close-enforcement guards into their own module
Part of #4013's module-split sequence (step 5, after transient-locks.ts,
signal-snapshot.ts, duplicate-detection.ts, and slop-detection.ts): the
5 review-evasion / one-shot-close guards (draft-dodge, reopen-reclose,
review-evasion self-close, review-evasion draft-conversion, repeated
draft-cycling) plus their shared scaffolding (withPrActuationLock,
evaluateCloseEnforcementGate, hasMaintainerOrOwnerPermission) move to
src/queue/review-evasion.ts. Pure mechanical move, no behavior change.

Only the 5 top-level entry points are exported (each called from
exactly one webhook-handler call site still in processors.ts); every
other function/type/constant here was already private with no caller
outside this cluster, so no re-export shim was needed -- nothing
external ever imported them from processors.ts.

Isolating this code surfaced two real, previously-invisible coverage
gaps:
- 8 best-effort `.catch(() => undefined)` handlers (label-application,
  review-tracking cleanup, courtesy comments, audit writes) had no v8
  ignore marker, unlike their identically-shaped siblings elsewhere in
  this same cluster -- added the same marker, matching this codebase's
  established convention for this exact class of fail-safe catch.
- The bot's-own-reopen exemption in the one-shot-reopen guard had no
  test at all -- added one exercising the real behavioral branch
  rather than suppressing it.
